requirements.txt
asgiref==3.2.8
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.4.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-searchable-select==1.5.0
djangorestframework==3.11.0
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
numpy==1.18.5
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.23.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
table==0.0.4
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
